requirements.txt
pandas
sqlalchemy
openpyxl
requests
beautifulsoup4
google-generativeai
pyyaml
psycopg2-binary
python-dotenv
schedule
google-generativeai